docker: Switch to node based image instead of texlive
If we utilize node's trixie based image, we have Python preinstalled
inside the Docker image. We can then install mermaid CLI immediately,
and then install the handful of other packages Sphinx depends on.

Additionally, switch to drawio version 27.0.9. The newer one has some
bugs in our existing drawio drawings that causes it to export an image
wrapped inside of a PDF instead of a vector image.